fix in status definition for diner request

+++ models/DinerRequest.js
@@ -29,7 +29,7 @@ module.exports = function(sequelize, DataTypes) {
       allowNull: false
     },
     status: {
-      type: DataTypes.INTEGER(3),
+      type: DataTypes.INTEGER(2),
       allowNull: false
     }
   }, {
